+++ lib/generation/column_generation.ts
@@ -1,4 +1,4 @@
-import {Aggregate, Alias, ColumnOperation, Get} from '../column_operations'
+import {Aggregate, Alias, ColumnOperation, Count, Get} from '../column_operations'
 import {joinWithCommaWhitespace} from '../parsing/javascript_parsing'
 import * as toSnakeCase from 'js-snakecase'
 
@@ -6,12 +6,16 @@ export function generateGet(get: Get): string {
     return `t${get.table}.${toSnakeCase(get.column)}`
 }
 
-export function generateAggregate(aggregate: Aggregate): string {
+function generateAggregate(aggregate: Aggregate): string {
     return `${aggregate.aggregation.toUpperCase()}(${generateGet(aggregate.get)})`
 }
 
-export function generateAlias(alias: Alias): string {
+function generateAlias(alias: Alias): string {
     return `${generateColumnOperation(alias.operation)} AS ${alias.alias}`
+}
+
+function generateCount(operation: Count): string {
+    return 'COUNT(*)';
 }
 
 function generateColumnOperation(operation: ColumnOperation): string {
@@ -22,6 +26,8 @@ function generateColumnOperation(operation: ColumnOperation): string {
             return generateGet(operation)
         case 'alias':
             return generateAlias(operation)
+        case 'count':
+            return generateCount(operation)
     }
 }